propagate onnx/onnx#8149 internal version type widening

+++ onnx_light/onnx_lib/common/ir_pb_converter.cc
@@ -23,7 +23,7 @@ static inline std::string to_std_string(const utils::String &s) {
 
 // Part 1: convert ONNX Protobuf to IR
 static std::unique_ptr<Graph> graphProtoToGraph(const GraphProto &gp, bool nested,
-                                                int ir_version = IR_VERSION);
+                                                int64_t ir_version = IR_VERSION);
 
 static Tensor tensorProtoToTensor(const TensorProto &tp) {
   Tensor ret;
@@ -129,7 +129,8 @@ static Tensor tensorProtoToTensor(const TensorProto &tp) {
   return ret;
 }
 
-static void convertAttribute(const AttributeProto &ap, Node &n, const int ir_version = IR_VERSION) {
+static void convertAttribute(const AttributeProto &ap, Node &n,
+                             const int64_t ir_version = IR_VERSION) {
   Symbol sym = Symbol(to_std_string(ap.name()));
   switch (ap.type()) {
   case AttributeProto::AttributeType::FLOAT:
@@ -214,7 +215,7 @@ static void convertAttribute(const AttributeProto &ap, Node &n, const int ir_ver
   }
 }
 
-static void convertAttributes(const NodeProto &np, Node &n, const int ir_version = IR_VERSION) {
+static void convertAttributes(const NodeProto &np, Node &n, const int64_t ir_version = IR_VERSION) {
   for (int i = 0; i < np.attribute().size(); i++) {
     convertAttribute(np.attribute()[i], n, ir_version);
   }
@@ -250,7 +251,8 @@ static Value *createDummyValue(const std::unique_ptr<Graph> &g, const std::strin
   return v;
 }
 
-std::unique_ptr<Graph> graphProtoToGraph(const GraphProto &gp, bool nested, const int ir_version) {
+std::unique_ptr<Graph> graphProtoToGraph(const GraphProto &gp, bool nested,
+                                         const int64_t ir_version) {
   auto g = std::make_unique<Graph>();
 
   if (gp.has_name()) {
